Move simple backend to lib/services/simple
